リファクタリング FavViewController
